requirements.txt
PyYAML==5.3.1
django==3.2.5
djangorestframework==3.12.4
cython==0.29.24
psycopg2-binary==2.9.1